Cleaning up some stuff from when the app was originally set up with example content. Reorganizing the other components into component folders.
