+++ src/components/button/button.scss
@@ -1,5 +1,5 @@
-@import '../variables';
-@import '../utilities';
+@import '../../assets/scss/variables';
+@import '../../assets/scss/utilities';
 
 .bttn {
   @include bttn;
