+++ src/components/headline/headline.scss
@@ -1,4 +1,4 @@
-@import "../variables";
+@import "../../assets/scss/variables";
 @import "headline-mixins";
 
 .headline {
